Update docs/configuring-playbook-client-schildichat-web.md: remove warning notification about lagging versions
As the SchildiChat Web docker image was updated with 14a3a01f43226487262d2e4c5d0a939ee93790be thanks to Aine, the warning is no longer relevant. Note that though the package name says it is a test, it is a test release for SchildiChat Web, and the codebase is Element Web v1.11.86.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>