Update docs/configuring-playbook-bot-draupnir.md: invite the Draupnir bot after installation
Now that the bot user for Draupnir is created automatically and you no longer need to register it manually since its 2.0.0 version, it does not seem to be sensible to synchronize descriptions about inviting bots between documents for Draupnir and Mjolnir. It is not friendly to instruct to invite the bot which does not exist just yet, only to let the known error message about the nonexistent user displayed.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>